display: Mode's rate use float64 type instead of uint16 solve same rate problem

+++ display/output.go
@@ -9,7 +9,7 @@ type Mode struct {
 	ID     uint32
 	Width  uint16
 	Height uint16
-	Rate   uint16
+	Rate   float64
 }
 type Output struct {
 	bestMode  randr.Mode
